Fixes #24 Fixed Auth test
